test(hmpb): split fixture generation tests
This should improve the performance of these tests by allowing them to
run in parallel.
